// src/api/generated/model/mgmtSoftwareModuleType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { Links } from './links';

export interface MgmtSoftwareModuleType {
  /** Entity was originally created by (User, AMQP-Controller, anonymous etc.) */
  readonly createdBy?: string;
  /** Entity was originally created at (timestamp UTC in milliseconds) */
  readonly createdAt?: number;
  /** Entity was last modified by (User, AMQP-Controller, anonymous etc.) */
  readonly lastModifiedBy?: string;
  /** Entity was last modified at (timestamp UTC in milliseconds) */
  readonly lastModifiedAt?: number;
  name: string;
  description?: string;
  /** Key that can be interpreted by the target */
  key: string;
  /** Colour assigned to the entity that could be used for representation purposes */
  colour?: string;
  /** Deleted flag, used for soft deleted entities */
  deleted?: boolean;
  /** The technical identifier of the entity */
  id: number;
  /** Software modules of that type can be assigned at this maximum number (e.g. operating system only once) */
  maxAssignments?: number;
  _links?: Links;
}